fix: Trending Products image bug
