+++ src/components/TrendingProducts/SingleProduct.module.scss
@@ -20,9 +20,10 @@
     width: 100%;
     height: 100%;
     position: relative;
+    overflow: hidden;
     img {
       max-width: 100%;
-      height: 100%;
+      max-height: 100%;
     }
   }
