+++ src/pages/OrderTracking/OrderTracking.js
@@ -1,77 +1,73 @@
 import React, { useEffect, useState } from "react";
 import { useSelector } from "react-redux";
+import { useParams } from "react-router-dom";
 import PageLayout from "../../containers/PageLayout";
 import PageContainer from "../../containers/PageContainer";
-import { Accordion, AccordionSummary, AccordionDetails } from "@mui/material";
+import Loader from "../../components/UI/Loader";
 import classes from "./OrderTracking.module.scss";
 
-import ExpandMoreIcon from "@mui/icons-material/ExpandMore";
 import ProgressBar from "../../components/ProgressBar";
 import OrderLog from "../../components/OrderLog";
+import CartItems from "../../components/CartItems/CartItems";
 
 const OrderTracking = () => {
   const myOrders = useSelector((store) => store.ordersReducer.myOrders);
-  const [keyword, setKeyword] = useState("");
+  const allProducts = useSelector((store) => store.productsReducer.products);
+  const currency = useSelector((store) => store.productsReducer.currency);
+  const [order, setOrder] = useState(null);
+  const params = useParams();
 
-  const filteredOrders = [...myOrders].filter((order) => {
-    if (keyword === "") {
-      return true;
-    } else {
-      return order.order_number.includes(keyword);
-    }
+  const orderNum = params.orderId;
+  const orderedProducts = order?.ordered_product.map((prod) => {
+    const product = { ...allProducts.find((el) => el.itemId === prod.itemId) };
+    product.totalPrice = prod.quantity * product.price;
+    return product;
   });
 
-  const markup = filteredOrders.reverse().map((order) => {
-    const orderIsCompleted = order.order_status === "delivered";
-    const date = new Date(order.order_estimation.seconds * 1000);
-    const EST = new Intl.DateTimeFormat("en-UK").format(date);
+  useEffect(() => {
+    setOrder((prevState) => {
+      return [...myOrders]?.find((order) => order?.order_number === orderNum);
+    });
+  }, [myOrders]);
 
-    return (
-      <Accordion>
-        <AccordionSummary
-          expandIcon={<ExpandMoreIcon />}
-          aria-controls="panel1a-content"
-          id="panel1a-header"
-        >
+  const orderIsCompleted = order ? order.order_status === "delivered" : null;
+  const date = order ? new Date(order.order_estimation.seconds * 1000) : null;
+  const EST = order ? new Intl.DateTimeFormat("en-UK").format(date) : null;
+
+  const markup = (
+    <section className={classes.section}>
+      <div>
+        <div className={classes.title}>
           <h2 className={classes["order-number"]}>
-            ORDER <span>#{order.order_number}</span>
+            ORDER <span>#{order?.order_number}</span>
           </h2>
           <p className={classes["est-arrival"]}>
             {orderIsCompleted ? "Delivered" : `Est: ${EST}`}
           </p>
-        </AccordionSummary>
-        <AccordionDetails>
+        </div>
+        <div>
           <ProgressBar order={order} />
-          <OrderLog order={order} />
-        </AccordionDetails>
-      </Accordion>
-    );
-  });
-
-  const handleChange = (e) => {
-    setKeyword(e.target.value);
-  };
+          <div className={classes.details}>
+            {orderedProducts && (
+              <div className={classes["ordered-product"]}>
+                <h2 className={classes.heading}>Ordered Products</h2>
+                <CartItems
+                  cartItems={orderedProducts}
+                  showControls={false}
+                  currency={currency}
+                />
+              </div>
+            )}
+            <OrderLog order={order} />
+          </div>
+        </div>
+      </div>
+    </section>
+  );
 
   return (
     <PageLayout title="Orders Tracking">
-      <PageContainer>
-        <section className={classes.section}>
-          <div className={classes["section-head"]}>
-            <h2>My Orders</h2>
-            <input
-              onChange={(e) => handleChange(e)}
-              type="number"
-              min="0"
-              placeholder="Order Number"
-            />
-          </div>
-          {filteredOrders.length > 0 ? (
-            markup
-          ) : (
-            <div className={classes["no-orders"]}>Nothing found</div>
-          )}
-        </section>
-      </PageContainer>
+      <PageContainer>{order ? markup : <Loader />}</PageContainer>
     </PageLayout>
   );
 };
