+++ src/pages/OrderTracking/OrderTracking.module.scss
@@ -3,6 +3,15 @@
 .section {
   padding: 5rem 0;
 }
+
+.title {
+  height: 100px;
+  position: relative;
+  display: flex;
+  align-items: center;
+  justify-content: space-between;
+  margin-bottom: 3rem;
+}
 .order-number {
   font-family: $primary-font;
   color: $text-color;
@@ -13,42 +22,37 @@
 }
 
 .est-arrival {
-  position: absolute;
   font-family: $secondary-font;
   font-size: 0.9rem;
-  right: 50px;
+
   color: $text-color;
 }
 
-.section-head {
-  width: 100%;
-  display: flex;
-  justify-content: space-between;
-  align-items: center;
-  margin: 1rem 0;
-  font-family: $primary-font;
-  h2 {
+.details {
+  display: grid;
+  grid-template-columns: 1fr 1fr;
+  grid-gap: 20px;
+  background-color: #f6f5ff;
+  align-items: flex-start;
+  padding: 1rem;
+  margin: 2rem 0;
+}
+
+.ordered-product {
+  padding: 1rem;
+  padding-right: 3rem;
+  border-right: 2px solid #e1e1e4;
+  .heading {
+    text-align: center;
+    margin: 1rem 0 2rem 0;
+    font-family: $primary-font;
     color: $text-color;
   }
-
-  input {
-    padding: 1rem;
-    border: 2px solid #e7e6ef;
-    outline: none;
-  }
-
-  input::placeholder {
-    font-family: $secondary-font;
-  }
 }
 
-.no-orders {
-  display: flex;
-  font-family: $primary-font;
-  color: $text-color;
-  font-size: 2rem;
-  width: 100%;
-  height: 200px;
-  align-items: center;
-  justify-content: center;
+@media only screen and (max-width: 950px) {
+  .details {
+    grid-template-columns: 1fr;
+    grid-template-rows: 1fr 1fr;
+  }
 }
