--- src/pages/OrdersTracking/OrderTracking.module.scss
@@ -1,58 +1,0 @@
-@import "../../styles/css-variables.scss";
-
-.section {
-  padding: 5rem 0;
-}
-
-.title {
-  height: 100px;
-  position: relative;
-  display: flex;
-  align-items: center;
-  justify-content: space-between;
-  margin-bottom: 3rem;
-}
-.order-number {
-  font-family: $primary-font;
-  color: $text-color;
-  font-size: 1.1rem;
-  span {
-    color: $secondary-color;
-  }
-}
-
-.est-arrival {
-  font-family: $secondary-font;
-  font-size: 0.9rem;
-
-  color: $text-color;
-}
-
-.details {
-  display: grid;
-  grid-template-columns: 1fr 1fr;
-  grid-gap: 20px;
-  background-color: #f6f5ff;
-  align-items: flex-start;
-  padding: 1rem;
-  margin: 2rem 0;
-}
-
-.ordered-product {
-  padding: 1rem;
-  padding-right: 3rem;
-  border-right: 2px solid #e1e1e4;
-  .heading {
-    text-align: center;
-    margin: 1rem 0 2rem 0;
-    font-family: $primary-font;
-    color: $text-color;
-  }
-}
-
-@media only screen and (max-width: 950px) {
-  .details {
-    grid-template-columns: 1fr;
-    grid-template-rows: 1fr 1fr;
-  }
-}
