--- src/pages/OrdersTracking/index.js
@@ -1,2 +1,0 @@
-import OrderTracking from "./OrderTracking";
-export default OrderTracking;
